refactor(InterpPoint): move InterpPoint to self-contained header
