+++ src/CMakeLists.txt
@@ -8,6 +8,7 @@ set(fers_SRCS logging.cpp main.cpp platform.cpp radar_system.cpp target.cpp worl
 	object.h antenna_factory.h response.h noise_generators.h timing.h hdf5_export.h python_extension.h
 	interpolation_set.h dsp_filters.h multipath_surface.h antenna_pattern.h jones_vector.h polarization_matrix.cpp
 	polarization_matrix.h interpolation_filter.cpp interpolation_filter.h
+	interpolation_point.h
 )
 
 add_executable(fers ${fers_SRCS})
